+++ src/main/java/com/koddy/server/member/application/usecase/query/response/MenteePrivateProfile.java
@@ -1,6 +1,7 @@
 package com.koddy.server.member.application.usecase.query.response;
 
 import com.koddy.server.member.domain.model.mentee.Mentee;
+import com.koddy.server.member.domain.model.response.LanguageResponse;
 
 public record MenteePrivateProfile(
         long id,
