+++ src/main/java/com/koddy/server/member/application/usecase/query/response/ReservedSchedule.java
@@ -3,6 +3,8 @@ package com.koddy.server.member.application.usecase.query.response;
 import com.koddy.server.coffeechat.domain.model.CoffeeChat;
 import com.koddy.server.coffeechat.domain.model.Reservation;
 import com.koddy.server.member.domain.model.mentor.Mentor;
+import com.koddy.server.member.domain.model.response.MentoringPeriodResponse;
+import com.koddy.server.member.domain.model.response.ScheduleResponse;
 
 import java.time.LocalDateTime;
 import java.util.List;
