+++ src/main/java/com/koddy/server/member/domain/model/response/LanguageResponse.java
@@ -1,4 +1,4 @@
-package com.koddy.server.member.application.usecase.query.response;
+package com.koddy.server.member.domain.model.response;
 
 import com.koddy.server.member.domain.model.Language;
