+++ src/test/java/com/koddy/server/member/application/usecase/GetMemberPrivateProfileUseCaseTest.java
@@ -4,12 +4,12 @@ import com.koddy.server.common.UnitTest;
 import com.koddy.server.common.fixture.MentoringPeriodFixture;
 import com.koddy.server.member.application.usecase.query.response.MenteePrivateProfile;
 import com.koddy.server.member.application.usecase.query.response.MentorPrivateProfile;
-import com.koddy.server.member.application.usecase.query.response.ScheduleResponse;
 import com.koddy.server.member.domain.model.Language;
 import com.koddy.server.member.domain.model.mentee.Mentee;
 import com.koddy.server.member.domain.model.mentor.Mentor;
 import com.koddy.server.member.domain.model.mentor.MentoringPeriod;
 import com.koddy.server.member.domain.model.mentor.Timeline;
+import com.koddy.server.member.domain.model.response.ScheduleResponse;
 import com.koddy.server.member.domain.repository.MenteeRepository;
 import com.koddy.server.member.domain.repository.MentorRepository;
 import org.junit.jupiter.api.DisplayName;
