+++ src/test/java/com/koddy/server/member/application/usecase/GetReservedScheduleUseCaseTest.java
@@ -6,9 +6,9 @@ import com.koddy.server.coffeechat.domain.repository.CoffeeChatRepository;
 import com.koddy.server.common.UnitTest;
 import com.koddy.server.member.application.usecase.query.GetReservedSchedule;
 import com.koddy.server.member.application.usecase.query.response.ReservedSchedule;
-import com.koddy.server.member.application.usecase.query.response.ScheduleResponse;
 import com.koddy.server.member.domain.model.mentee.Mentee;
 import com.koddy.server.member.domain.model.mentor.Mentor;
+import com.koddy.server.member.domain.model.response.ScheduleResponse;
 import com.koddy.server.member.domain.repository.MentorRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
